proof_of_knowledge/requirements.txt
# black==19.10b0
eth-brownie>=1.11.0,<2.0.0
dotmap==1.3.23
python-dotenv==0.16.0
# tabulate==0.8.7
# rich==9.3.0
